perf: set cmd to footer
