+++ proxy.conf.json
@@ -12,12 +12,12 @@
     "changeOrigin": true
   },
   "/koko/static/*": {
-    "target": "http://localhost:5000",
+    "target": "http://localhost:5050",
     "secure": false,
     "ws": true
   },
   "/koko/*": {
-    "target": "http://localhost:5000",
+    "target": "http://localhost:5050",
     "secure": false,
     "ws": true,
     "changeOrigin": true
