+++ src/app/elements/content/content.component.scss
@@ -1,5 +1,4 @@
 $-border-x: #5a5959;
-$-border-y: var(--el-main-bg-color);
 
 #content {
   height: calc(100% - 30px);
@@ -63,7 +62,13 @@ $-border-y: var(--el-main-bg-color);
   }
 
   #winContainer {
+    display: flex;
+    flex-direction: column;
     height: 100%;
+
+    elements-content-window {
+      flex: 1;
+    }
 
     .content-placeholder {
       align-items: center;
@@ -97,10 +102,6 @@ $-border-y: var(--el-main-bg-color);
         }
       }
     }
-
-    &.batch-input {
-      height: calc(100% - 30px);
-    }
   }
 }
 
@@ -118,161 +119,6 @@ $-border-y: var(--el-main-bg-color);
 
 .tabs::-webkit-scrollbar-thumb {
   background-color: #F5F5F5;
-}
-
-#batchCommandDiv {
-  display: flex;
-  height: 29px;
-  border-left: solid 1px $-border-y;
-  width: 100%;
-  padding: 0 2px 0 0;
-  background: var(--el-main-bg-color);
-  border-top: solid 1px rgba(255, 255, 255, 0.2);
-
-  &:focus-visible {
-    outline: none;
-  }
-
-  i {
-    margin: auto 7px;
-    color: #ffffff;
-    cursor: pointer;
-
-    &:hover {
-      color: #d6cbcb;
-    }
-  }
-
-  .view_icon {
-    width: 12px;
-    height: 12px;
-    margin: 0 !important;
-    margin-right: 3px !important;
-    vertical-align: middle;
-    display: inline-block;
-  }
-
-  .view_icon::before {
-    font-family: FontAwesome;
-    color: #ffffff;
-    display: inline-block;
-  }
-}
-
-#batchCommandDiv input {
-  flex: 1;
-  height: 29px;
-  width: 100%;
-  line-height: 29px;
-  border: none;
-  background: var(--el-main-bg-color);
-  color: #ffffff;
-
-  &:focus-visible {
-    outline: none;
-  }
-}
-
-#batchCommandDiv {
-  .send-range {
-    width: auto;
-    margin-left: 5px;
-
-    &::ng-deep .mat-select-trigger {
-      min-width: 70px;
-      line-height: 29px;
-
-      .mat-select-value {
-        max-width: none;
-        color: #ffffff;
-      }
-
-      .mat-select-arrow {
-        color: #ffffff;
-      }
-    }
-  }
-
-  .command-list {
-    width: 100%;
-    height: 100%;
-    margin: 0;
-    padding: 3px 0;
-    overflow: hidden;
-    white-space: nowrap;
-    transition: all .3s;
-
-    .command-box {
-      display: inline-block;
-      max-width: 260px;
-      height: 100%;
-      line-height: 15px;
-      padding: 4px;
-      margin-right: 5px;
-      border-radius: 4px;
-      overflow: hidden;
-      white-space: nowrap;
-      text-overflow: ellipsis;
-      cursor: pointer;
-      user-select: none;
-      color: #d6cbcb;
-      background-color: #463e3e;
-      &:hover {
-        color: #ffffff;
-      }
-
-      &:active {
-        color: #ccc8c8;
-      }
-    }
-    .command-box-variable {
-      @extend .command-box;
-      background-color: #6f4f3a;
-    }
-  }
-
-  .not-command {
-    flex: 1;
-    line-height: 29px;
-    padding-left: 6px;
-    color: #717171;
-  }
-
-  .action-left,
-  .action-right {
-    line-height: 29px;
-  }
-
-  .action-right {
-    display: flex;
-  }
-}
-
-::ng-deep .command-panel.mat-select-panel {
-  margin-top: 27px;
-  min-width: 100% !important;
-  border-radius: initial !important;
-  background: black;
-  padding: 10px 0 !important;
-}
-
-::ng-deep .command-panel .mat-option {
-  color: #f0f0f1;
-
-  &:hover {
-    background-color: rgba(53, 54, 51, 0.7) !important;
-  }
-}
-
-::ng-deep .command-panel .mat-option.mat-active {
-  color: #f0f0f1;
-  font-weight: bold;
-  background-color: rgba(53, 54, 51, 0.8) !important;
-}
-
-.command-option {
-  line-height: 2.2em !important;
-  height: 2.2em !important;
 }
 
 .view-menu {
@@ -407,10 +253,4 @@ tr:hover {
 
 .view-menu__icon.fa {
   width: 15px;
-}
-
-.input-box {
-  display: flex;
-  width: 100%;
-  height: 100%;
 }
